Updates tts folder
